Added some sample tests and Action.yml

+++ AirplaneAPI/Database/DTO/FlightDTO.cs
@@ -9,9 +9,7 @@ public class FlightDTO
     public DateTime DepartureTime { get; set; } // Time of departure
     public DateTime ArrivalTime { get; set; } // Time of arrival
     public string FlightNumber { get; set; } // Flight number
-    
-    //This property is could be calculated from the airports' coordinates and would be redundant
-    public int Distance { get; set; } // Distance between the airports
+    public double Distance { get; set; } // Distance for the flight in km
     public int FuelCost { get; set; } // Cost of fuel
     
     public TimeSpan Duration => ArrivalTime - DepartureTime; // Duration of the flight
